refactor : 에러코드 관련 수정
